add pharokka proteins3

+++ src/phold/subcommands/compare.py
@@ -207,7 +207,8 @@ def subcommand_compare(
                     logger.info(
                         f"Checked that the 3Di CDS file {fasta_3di_input} exists from phold predict"
                     )
-                    shutil.copyfile(fasta_3di_input, fasta_3di)
+                    if fasta_3di_input != fasta_3di: # only copy the file if doesnt already exist (will be the case for run or proteins)
+                        shutil.copyfile(fasta_3di_input, fasta_3di)
                 else:
                     logger.error(
                         f"The 3Di CDS file {fasta_3di_input} does not exist. Please run phold predict and/or check the prediction directory {predictions_dir}"
@@ -217,7 +218,8 @@ def subcommand_compare(
                     logger.info(
                         f"Checked that the AA CDS file {fasta_aa_input} exists from phold predict."
                     )
-                    shutil.copyfile(fasta_aa_input, fasta_aa)
+                    if fasta_3di_input != fasta_3di: # only copy the file if doesnt already exist (will be the case for run or proteins)
+                        shutil.copyfile(fasta_3di_input, fasta_3di)
                 else:
                     logger.error(
                         f"The AA CDS file {fasta_aa_input} does not exist. Please run phold predict and/or check the prediction directory {predictions_dir}"
